+++ lib/kirby.php
@@ -397,10 +397,10 @@ class a {
     $direction = (strtolower($direction) == 'desc') ? SORT_DESC : SORT_ASC;
     $helper    = array();
 
-    foreach($array as $key => $row) {
+    foreach($array as $key => $row) {      
       $helper[$key] = (is_object($row)) ? (method_exists($row, $field)) ? str::lower($row->$field()) : str::lower($row->$field) : str::lower($row[$field]);
     }
-    
+
     // natural sorting    
     if($method === 'natural') {
       natsort($helper);